requirements.txt
fastapi==0.103.2
pydantic==2.3.0
uvicorn==0.24.0
openai-whisper==20230918
transformers==4.44.0
torch==2.1.0
numpy==1.22.0
python-multipart
ffmpeg-python
accelerate
ngrok
pydub